requirements.txt
Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
paramiko==3.4.0
docker==7.0.0